+++ src/anemoi/datasets/commands/recipe/dump.py
@@ -36,6 +36,9 @@ class Dumper:
     def start_end_dates(self, start, end, frequency):
         return repr((str(start), str(end), frequency_to_string(frequency)))
 
+    def steps(self, start, end, frequency):
+        return repr((frequency_to_string(start), frequency_to_string(end), frequency_to_string(frequency)))
+
 
 def dump_recipe(config: dict, dumper=None) -> str:
     recipe = Recipe(**config)